Updated Version 2

+++ Module 2/NoiseChannel.py
@@ -11,7 +11,7 @@ def NoiseChannel(inp_X1, inp_X2, inpSize, snrArray, snrSize):
 		noise = [0]*inpSize
 		l = []
 		for k in range(inpSize):
-			noise[k] = sqrt(10**((-snrArray[i])/10.0)) * random.randn(1,1)
+			noise[k] = float(sqrt(10**((-snrArray[i])/10.0)) * random.randn(1,1))
 			noise[k] *= (2**8)
 			l.append(inp_X1[k]+noise[k])
 		noisedSignal_X1.append(l)
@@ -20,9 +20,9 @@ def NoiseChannel(inp_X1, inp_X2, inpSize, snrArray, snrSize):
 		noise = [0]*inpSize
 		l = []
 		for k in range(inpSize):
-			noise[k] = sqrt(10**((-snrArray[i])/10.0)) * random.randn(1,1)
+			noise[k] = float(sqrt(10**((-snrArray[i])/10.0)) * random.randn(1,1))
 			noise[k] *= (2**8)
 			l.append(inp_X2[k]+noise[k])
 		noisedSignal_X2.append(l)
-
+	
 	return noisedSignal_X1, noisedSignal_X2
